feat(settings): 拆分 Models 设置页并移植 AGENTS.md / APPEND_SYSTEM.md 编辑器
settings-config.js 曾是一个 1600 行的巨型模块，违反单一职责的模块纪律；
同时 feature-v3 配置页的 AGENTS.md 与 APPEND_SYSTEM.md（追加系统提示）
编辑器在拆分中缺失。本次：

- 将 provider/API-keys/models.json 管理整体迁出为 models-page.js
  （Models 标签页），settings-config.js 只保留 agent settings.json 编辑器
- 按迁移 playbook 从 feature-v3 逐字移植 AGENTS.md / APPEND_SYSTEM.md
  两个 markdown 编辑器进「高级配置」标签页（wireFileEditor 工厂 +
  HTML 段落 + 四语言文案 + 编辑器测试），仅做 4 项声明的适配
- 扩展 picot-config 增加 readTextFile/writeTextFile 与 4 个纯文本 ops
  （缺文件读空、无 JSON 校验），与既有 config ops 契约一致
- 空模型下拉框的「打开设置」路由改为 models 标签页（app.js:2266）
- 移除随 Protection 标记删除的 toggle-auth 遗留代码
- 测试：settings 目录 17 文件 120 用例 + 扩展 15 用例全绿；
  bun run check 无新增违规
