+++ public/locales/es.json
@@ -103,7 +103,7 @@
   "models": {
     "searchPlaceholder": "Buscar modelos…",
     "emptyTitle": "No hay modelos disponibles",
-    "emptyHelp": "No hay claves API configuradas. Establece una clave en Ajustes → Proveedores.",
+    "emptyHelp": "No hay claves API configuradas. Establece una clave en Ajustes → Modelos.",
     "openSettings": "Abrir ajustes"
   },
   "queue": {
@@ -578,7 +578,7 @@
     "general": "General",
     "extensions": "Extensiones",
     "usage": "Uso",
-    "configuration": "Proveedores",
+    "configuration": "Configuración avanzada",
     "agentInbox": "Bandeja del agente",
     "backToChat": "Volver al chat",
     "appearance": "Apariencia",
@@ -602,6 +602,10 @@
     "protection": "Protección",
     "requireLogin": "Requerir inicio de sesión",
     "agentConfigFile": "Archivo de configuración del agente",
+    "agentsMd": "AGENTS.md",
+    "agentsContextFile": "Archivo de contexto del agente (AGENTS.md)",
+    "appendSystemMd": "APPEND_SYSTEM.md",
+    "appendSystemPromptFile": "Archivo de anexo al prompt del sistema (APPEND_SYSTEM.md)",
     "providersFile": "Archivo de proveedores",
     "llmProviders": "Proveedores LLM",
     "browseCommunityPackages": "Explorar paquetes de la comunidad",
@@ -640,6 +644,7 @@
       "configured": "Configurado"
     },
     "models": {
+      "title": "Modelos",
       "helpPre": "Edita",
       "helpPost1": "para añadir proveedores personalizados (Ollama, vLLM, LM Studio, proxies compatibles con OpenAI, anulaciones de enrutamiento de OpenRouter, etc). Consulta la",
       "docsLink": "documentación de models.json",
