+++ public/locales/ja.json
@@ -103,7 +103,7 @@
   "models": {
     "searchPlaceholder": "モデルを検索…",
     "emptyTitle": "利用可能なモデルがありません",
-    "emptyHelp": "APIキーが設定されていません。設定 → プロバイダーでキーを設定してください。",
+    "emptyHelp": "APIキーが設定されていません。設定 → モデル でキーを設定してください。",
     "openSettings": "設定を開く"
   },
   "queue": {
@@ -578,7 +578,7 @@
     "general": "一般",
     "extensions": "拡張機能",
     "usage": "使用量",
-    "configuration": "プロバイダー",
+    "configuration": "詳細設定",
     "agentInbox": "エージェント受信箱",
     "backToChat": "チャットに戻る",
     "appearance": "外観",
@@ -602,6 +602,10 @@
     "protection": "保護",
     "requireLogin": "ログインを要求",
     "agentConfigFile": "エージェント設定ファイル",
+    "agentsMd": "AGENTS.md",
+    "agentsContextFile": "エージェントコンテキストファイル（AGENTS.md）",
+    "appendSystemMd": "APPEND_SYSTEM.md",
+    "appendSystemPromptFile": "システムプロンプト追記ファイル（APPEND_SYSTEM.md）",
     "providersFile": "プロバイダーファイル",
     "llmProviders": "LLMプロバイダー",
     "browseCommunityPackages": "コミュニティパッケージを閲覧",
@@ -640,6 +644,7 @@
       "configured": "設定済み"
     },
     "models": {
+      "title": "モデル",
       "helpPre": "編集",
       "helpPost1": "でカスタムプロバイダーを追加できます（Ollama、vLLM、LM Studio、OpenAI互換プロキシ、OpenRouterルーティングオーバーライドなど）。完全なスキーマは",
       "docsLink": "models.jsonドキュメント",
